move variant methods for sign in gate to own file
